Use tabular numbers in timer fields if available
This is only relevant if you have changed the default timer font to something other than 'monospace' and the selected font supports this functionality. It will prevent the position of the digits in the timer from 'jumping' around while an exam is in progress, which can be distracting for students.